Upload of the needed files to execute the demo.
There are two versions:
tfg_project_full.m : includes the Color Deconvolution (which needs the file Colour_Deconvolution2.m).
tfg_project_no_conv.m : It understands that the Color deconvolution has been performed previously.

In both cases, the folder path to the histological images has to be changed to match your own. 
In my case, the path is "images/selection" for the histological images, and "images/selection_deconv" for  the pre deconvoluted images.